add with_* methods into selectable trait.
